EDIT colors for higher contrast
Customize the colors to suit my preferences. The default colors seem
undersaturated, resulting in low contrast, and hence is a bit hard on my
eyes. Not that I'm all good at picking colors, so I just started from
the primary color versions and went darker/lighter as needed.

Also added a new diff background color, as it was previously using the
visual selection color. This means we can now see what we're
highlighting when we work with diffs.